+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-# Change Log
-- 2026-07-11T21:28:44.838Z: Pulse 68620 - Evolutionary growth detected.
+# Evolution Log
+- 2026-07-11T21:29:49.851Z: Pulse 68630 - Integrity Stable.